+++ src/dp_emission.h
@@ -10,7 +10,7 @@ struct seq_code;
 
 struct dp_emission
 {
-    double*   score;  /**< Sequence emission score of a state. */
+    float*    score;  /**< Sequence emission score of a state. */
     uint32_t* offset; /**< Maps state to score array offset. */
 };
 
@@ -19,13 +19,13 @@ struct dp_emission const* dp_emission_create(struct seq_code const*      seq_cod
 void                      dp_emission_destroy(struct dp_emission const* emission);
 void dp_emission_offsets_dump(struct dp_emission const* emission, uint32_t nstates);
 struct dp_emission const* dp_emission_read(FILE* stream);
-static inline double      dp_emission_score(struct dp_emission const* emission, uint32_t state,
+static inline float       dp_emission_score(struct dp_emission const* emission, uint32_t state,
                                             unsigned seq_code);
 void dp_emission_scores_dump(struct dp_emission const* emission, uint32_t nstates);
 int  dp_emission_write(struct dp_emission const* emission, uint32_t nstates, FILE* stream);
 
-static inline double dp_emission_score(struct dp_emission const* emission, uint32_t state,
-                                       unsigned seq_code)
+static inline float dp_emission_score(struct dp_emission const* emission, uint32_t state,
+                                      unsigned seq_code)
 {
     return emission->score[emission->offset[state] + seq_code];
 }
